+++ changelog.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Checks if CommandBox is installed, if not, it asks to install it for you.
+- `jq` dependency check
+
 ## [1.6.0] - 2025-06-05
 
 ### Added